+++ library/Kathu/Graphics/Camera.hs
@@ -8,4 +8,7 @@ import GHC.Generics
 
 newtype Camera = Camera {zoom :: Double} deriving (Show, Eq, Generic)
 
+defaultCamera :: Camera
+defaultCamera = Camera 1
+
 instance Component Camera where type Storage Camera = Unique Camera